# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subway Guatemala Customer API",
        "description": "API REST para la aplicación móvil de clientes de Subway Guatemala. Incluye autenticación multi-canal (email/password, Google OAuth, Apple Sign-In), gestión de perfil, dispositivos FCM y futuras funcionalidades de pedidos y lealtad.",
        "contact": {
            "email": "dev@subwayguatemala.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Application Server"
        }
    ],
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new customer",
                "description": "Creates a new customer account with email and password. Sends email verification notification.",
                "operationId": "7ca6fc7989202054d29707fef91a2d4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Customer full name",
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "description": "Valid email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "description": "Min 8 characters, 1 uppercase, 1 number, 1 special char",
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePass123!"
                                    },
                                    "password_confirmation": {
                                        "description": "Must match password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePass123!"
                                    },
                                    "phone": {
                                        "description": "Optional phone number",
                                        "type": "string",
                                        "example": "+50212345678",
                                        "nullable": true
                                    },
                                    "birth_date": {
                                        "description": "Optional birth date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15",
                                        "nullable": true
                                    },
                                    "gender": {
                                        "description": "Optional gender",
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female",
                                            "other"
                                        ],
                                        "example": "male",
                                        "nullable": true
                                    },
                                    "device_name": {
                                        "description": "Device name for token",
                                        "type": "string",
                                        "example": "iPhone 14 Pro",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Registro exitoso. Por favor verifica tu email."
                                        },
                                        "data": {
                                            "properties": {
                                                "customer": {
                                                    "$ref": "#/components/schemas/Customer"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123xyz..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login with email and password",
                "description": "Authenticates customer and returns Sanctum token. Rate limited to 5 attempts per minute.",
                "operationId": "9d0bb5c770cbec19b0b9566606ef4adb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "SecurePass123!"
                                    },
                                    "device_name": {
                                        "description": "Device name for token",
                                        "type": "string",
                                        "example": "iPhone 14 Pro",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inicio de sesión exitoso."
                                        },
                                        "data": {
                                            "properties": {
                                                "customer": {
                                                    "$ref": "#/components/schemas/Customer"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123xyz..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials or OAuth account"
                    },
                    "429": {
                        "description": "Too many login attempts"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from current device",
                "description": "Revokes the current access token. Other devices remain logged in.",
                "operationId": "3a27a9e02ceb521eb3f98c8d2394aef9",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout-all": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout from all devices",
                "description": "Revokes all access tokens. Customer will be logged out from all devices.",
                "operationId": "afe5fcc3199730bf8caf978b67c6a0c0",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se cerraron todas las sesiones exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Obtain from /api/v1/auth/login, /api/v1/auth/register, or OAuth endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}